move initialize command
